+++ packages/terra-overlay/CHANGELOG.md
@@ -1,10 +1,16 @@
 Changelog
 =========
+### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 
 Unreleased
 ----------
+
+3.20.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies
+* Updated jest snapshots
 
 ### Changed
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`